smp: better identification of lapic

+++ kernel/src/cpu/smp.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <misc/utils.h>
 
-uint16_t smpID();
+uint8_t smpID();
 uint16_t smpGetCores();
 void smpBootstrap();
